requirements.txt
Django==5.1.7
psycopg2-binary==2.9.9
gunicorn==21.2.0
djangorestframework==3.15.1
django-filter==24.2
whitenoise==6.6.0
